script/requirements.txt
pandas
pytest
sqlalchemy
python-logging-loki
psycopg2-binary